pickynes
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@3810 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/mod/say/mod_say_en/mod_say_en.c
@@ -38,7 +38,7 @@ static const char modname[] = "mod_say_en";
 #define say_num(num) {\
 		char tmp[80];\
 		switch_status_t status;\
-		snprintf(tmp, sizeof(tmp), "%d", num);\
+		snprintf(tmp, sizeof(tmp), "%u", (unsigned)num);				\
 	if ((status = en_say_general_count(session, tmp, SST_ITEMS, SSM_PRONOUNCED, input_callback, buf, buflen)) != SWITCH_STATUS_SUCCESS) {\
 		return status;\
 	}}\
